requirements.txt
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy==1.10.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
stripe
drf-yasg
celery
redis
django-celery-beat
